done with swagger

+++ src/main/java/com/helloguyzs/syfetask/dto/goals/UpdateGoalRequest.java
@@ -7,6 +7,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -15,7 +16,7 @@ import java.time.LocalDate;
 public class UpdateGoalRequest {
 
     @Min(value=1 , message = "Target amount must be greater than 0")
-    private Double targetAmount;
+    private BigDecimal targetAmount;
 
     @Future(message = "Target date must be a future date")
     private LocalDate targetDate;
